Conflicts:
	src/com/luxoft/bankapp/application/BankApplication.java
	src/com/luxoft/bankapp/application/BankClientRemoteOffice.java
	src/com/luxoft/bankapp/application/BankServer.java
	src/com/luxoft/bankapp/application/BankUI.java
	src/com/luxoft/bankapp/commander/Commander.java
	src/com/luxoft/bankapp/commander/command/BankInfoCommand.java
	src/com/luxoft/bankapp/commander/command/FindClientCommand.java
	src/com/luxoft/bankapp/commander/command/GetActiveAccountCommand.java
	src/com/luxoft/bankapp/commander/command/LoadCommand.java
	src/com/luxoft/bankapp/commander/command/SaveCommand.java
	src/com/luxoft/bankapp/commander/command/WithdrawCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/AddAccountCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/AddClientCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/BankFeedCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/DepositCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/SetActiveAccountCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/TransferCommand.java
	src/com/luxoft/bankapp/model/impl/Bank.java
	src/com/luxoft/bankapp/model/impl/Client.java

+++ src/com/luxoft/bankapp/dao/impl/ClientDaoImpl.java
@@ -2,28 +2,21 @@ package com.luxoft.bankapp.dao.impl;
 
 import com.luxoft.bankapp.dao.AccountDao;
 import com.luxoft.bankapp.dao.BaseDao;
+import com.luxoft.bankapp.dao.ClientDao;
 import com.luxoft.bankapp.dao.exceptions.DaoException;
 import com.luxoft.bankapp.model.Account;
 import com.luxoft.bankapp.model.Gender;
 import com.luxoft.bankapp.model.exceptions.ClientNotExistsException;
 import com.luxoft.bankapp.model.impl.Bank;
-import com.luxoft.bankapp.model.impl.CheckingAccount;
 import com.luxoft.bankapp.model.impl.Client;
-import com.luxoft.bankapp.dao.ClientDao;
-import com.luxoft.bankapp.model.impl.SavingAccount;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-/**
- * Created by SCJP on 27.01.15.
- */
+
 public class ClientDaoImpl implements ClientDao {
 
     private BaseDao baseDao = new BaseDaoImpl();
@@ -38,7 +31,7 @@ public class ClientDaoImpl implements ClientDao {
             preparedStatement.setLong(1, bank.getId());
             preparedStatement.setString(2, name);
 
-            if(!preparedStatement.execute()){
+            if (!preparedStatement.execute()) {
                 throw new DaoException("impossible to find the client in db.");
             }
             ResultSet rs = preparedStatement.getResultSet();
@@ -53,9 +46,9 @@ public class ClientDaoImpl implements ClientDao {
             long idActiveAccount = rs.getLong(8);
 
             Gender gender;
-            if(genderLetter.equals("m")){
+            if (genderLetter.equals("m")) {
                 gender = Gender.MALE;
-            } else if(genderLetter.equals("f")){
+            } else if (genderLetter.equals("f")) {
                 gender = Gender.FEMALE;
             } else {
                 throw new DaoException("incorrect data in db, impossible to load the client");
@@ -64,7 +57,13 @@ public class ClientDaoImpl implements ClientDao {
             AccountDao accountDao = new AccountDaoImpl();
             Set<Account> accounts = new HashSet<>(accountDao.getClientAccounts(idClient));
 
-            Account activeAccount = accountDao.findAccountById(idActiveAccount);
+            //Account activeAccount = accountDao.findAccountById(idActiveAccount);
+            Account activeAccount = null;
+            for (Account account : accounts) {
+                if (account.getId() == idActiveAccount) {
+                    activeAccount = account;
+                }
+            }
 
             client = new Client();
             client.setId(idClient);
@@ -87,7 +86,7 @@ public class ClientDaoImpl implements ClientDao {
     }
 
     @Override
-    public Client findClientById(int id) throws ClientNotExistsException, DaoException {
+    public Client findClientById(long id) throws ClientNotExistsException, DaoException {
         Connection conn = baseDao.openConnection();
         String sql = "select * from clients where id = (?)";
         Client client = null;
@@ -95,43 +94,45 @@ public class ClientDaoImpl implements ClientDao {
             final PreparedStatement preparedStatement = conn.prepareStatement(sql);
             preparedStatement.setLong(1, id);
 
-            if(!preparedStatement.execute()){
+            if (!preparedStatement.execute()) {
                 throw new DaoException("impossible to find the client in db.");
             }
             ResultSet rs = preparedStatement.getResultSet();
-            rs.next();
-            long idClient = rs.getLong(1);
-            String clientName = rs.getString(2);
-            String city = rs.getString(3);
-            String email = rs.getString(4);
-            String phone = rs.getString(5);
-            float overdraft = rs.getFloat(6);
-            String genderLetter = rs.getString(7);
-            long idActiveAccount = rs.getLong(8);
+            if (rs.next()) {
+                long idClient = rs.getLong(1);
+                String clientName = rs.getString(2);
+                String city = rs.getString(3);
+                String email = rs.getString(4);
+                String phone = rs.getString(5);
+                float overdraft = rs.getFloat(6);
+                String genderLetter = rs.getString(7);
+                long idActiveAccount = rs.getLong(8);
 
-            Gender gender;
-            if(genderLetter.equals("m")){
-                gender = Gender.MALE;
-            } else if(genderLetter.equals("f")){
-                gender = Gender.FEMALE;
-            } else {
-                throw new DaoException("incorrect data in db, impossible to load the client");
+                Gender gender;
+                if (genderLetter.equals("m")) {
+                    gender = Gender.MALE;
+                } else if (genderLetter.equals("f")) {
+                    gender = Gender.FEMALE;
+                } else {
+                    throw new DaoException("incorrect data in db, impossible to load the client");
+                }
+
+                AccountDao accountDao = new AccountDaoImpl();
+                Set<Account> accounts = new HashSet<>(accountDao.getClientAccounts(idClient));
+
+                Account activeAccount = accountDao.findAccountById(idActiveAccount);
+
+                client = new Client();
+                client.setId(idClient);
+                client.setName(clientName);
+                client.setCity(city);
+                client.setEmail(email);
+                client.setPhone(phone);
+                client.setOverdraft(overdraft);
+                client.setGender(gender);
+                client.setAccounts(accounts);
+                client.setActiveAccount(activeAccount);
             }
-
-            AccountDao accountDao = new AccountDaoImpl();
-            Set<Account> accounts = new HashSet<>(accountDao.getClientAccounts(idClient));
-
-            Account activeAccount = accountDao.findAccountById(idActiveAccount);
-
-            client.setId(idClient);
-            client.setName(clientName);
-            client.setCity(city);
-            client.setEmail(email);
-            client.setPhone(phone);
-            client.setOverdraft(overdraft);
-            client.setGender(gender);
-            client.setAccounts(accounts);
-            client.setActiveAccount(activeAccount);
 
             rs.close();
             preparedStatement.close();
@@ -144,19 +145,19 @@ public class ClientDaoImpl implements ClientDao {
 
     @Override
     public List<Client> getAllClients(Bank bank) throws DaoException {
-        List<Client>clients = null;
+        List<Client> clients = null;
         Connection conn = baseDao.openConnection();
         String sql = "select * from clients where id_bank = (?)";
         try {
             final PreparedStatement preparedStatement = conn.prepareStatement(sql);
             preparedStatement.setLong(1, bank.getId());
 
-            if(!preparedStatement.execute()){
+            if (!preparedStatement.execute()) {
                 throw new DaoException("impossible to get clients from db.");
             }
             ResultSet rs = preparedStatement.getResultSet();
             clients = new ArrayList<>();
-            while(rs.next()){
+            while (rs.next()) {
 
                 Client client = new Client();
                 long idClient = rs.getLong(1);
@@ -169,9 +170,9 @@ public class ClientDaoImpl implements ClientDao {
                 long idActiveAccount = rs.getLong(8);
 
                 Gender gender;
-                if(genderLetter.equals("m")){
+                if (genderLetter.equals("m")) {
                     gender = Gender.MALE;
-                } else if(genderLetter.equals("f")){
+                } else if (genderLetter.equals("f")) {
                     gender = Gender.FEMALE;
                 } else {
                     throw new DaoException("incorrect data in db, impossible to load the client");
@@ -203,11 +204,10 @@ public class ClientDaoImpl implements ClientDao {
         return clients;
     }
 
-    @Override
-    public void save(Client client) throws DaoException {
+    private void insert(Client client) throws DaoException {
         Connection conn = baseDao.openConnection();
-        String sql = "insert into clients (name, city, email, phone, overdraft, balance, gender)"+
-                " values (?, ?, ?, ?, ?, ?, ?, ?)";
+        String sql = "insert into clients (name, city, email, phone, overdraft, gender, id_bank)" +
+                " values (?, ?, ?, ?, ?, ?, 1)";
         try {
             final PreparedStatement preparedStatement = conn.prepareStatement(sql);
             preparedStatement.setString(1, client.getName());
@@ -215,20 +215,31 @@ public class ClientDaoImpl implements ClientDao {
             preparedStatement.setString(3, client.getEmail());
             preparedStatement.setString(4, client.getPhone());
             preparedStatement.setFloat(5, client.getOverdraft());
-            preparedStatement.setFloat(6, client.getBalance());
-            preparedStatement.setString(7, String.valueOf(client.getGender().getClass().getSimpleName().toLowerCase().charAt(0)));
+            preparedStatement.setString(6, String.valueOf(client.getGender().toString().toLowerCase().charAt(0)));
 
-            if(preparedStatement.executeUpdate() == 0){
+            if (preparedStatement.executeUpdate() == 0) {
                 throw new DaoException("impossible to save account in db. transaction is rolled back");
             }
             ResultSet rs = preparedStatement.getGeneratedKeys();
-            if(rs == null || !rs.next()){
+            if (rs == null || !rs.next()) {
                 throw new DaoException("impossible to save the client in db. transaction is rolled back");
             }
             Long id = rs.getLong(1);
             rs.close();
             preparedStatement.close();
             client.setId(id);
+
+            AccountDao accountDao = new AccountDaoImpl();
+            for (Account account : client.getAccounts()) {
+                accountDao.save(account, client);
+            }
+
+            String sqlSetActiveAccount = "update clients set id_active_account = (?) where id = (?)";
+            PreparedStatement setActiveAccountStatement = conn.prepareStatement(sqlSetActiveAccount);
+            setActiveAccountStatement.setLong(1, client.getActiveAccount().getId());
+            setActiveAccountStatement.setLong(2, client.getId());
+            setActiveAccountStatement.executeUpdate();
+            setActiveAccountStatement.close();
         } catch (SQLException e) {
             throw new DaoException(e.getMessage());
         }
@@ -236,14 +247,63 @@ public class ClientDaoImpl implements ClientDao {
     }
 
     @Override
+    public void save(Client client) throws DaoException {
+        if (client.getId() == -1) {
+            insert(client);
+        } else {
+            Connection conn = baseDao.openConnection();
+            String sql = "update clients set name = (?), city = (?), email = (?), phone = (?), overdraft = (?), gender = (?)" +
+                    " where id = (?)";
+            try {
+                final PreparedStatement preparedStatement = conn.prepareStatement(sql);
+                preparedStatement.setString(1, client.getName());
+                preparedStatement.setString(2, client.getCity());
+                preparedStatement.setString(3, client.getEmail());
+                preparedStatement.setString(4, client.getPhone());
+                preparedStatement.setFloat(5, client.getOverdraft());
+                preparedStatement.setString(6, String.valueOf(client.getGender().toString().toLowerCase().charAt(0)));
+                preparedStatement.setLong(7, client.getId());
+
+                if (preparedStatement.executeUpdate() == 0) {
+                    throw new DaoException("impossible to save account in db. transaction is rolled back");
+                }
+                preparedStatement.close();
+                AccountDao accountDao = new AccountDaoImpl();
+                for (Account account : client.getAccounts()) {
+                    accountDao.save(account, client);
+                }
+
+                String sqlSetActiveAccount = "update clients set id_active_account = (?) where id = (?)";
+                PreparedStatement setActiveAccountStatement = conn.prepareStatement(sqlSetActiveAccount);
+                setActiveAccountStatement.setLong(1, client.getActiveAccount().getId());
+                setActiveAccountStatement.setLong(2, client.getId());
+                setActiveAccountStatement.executeUpdate();
+                setActiveAccountStatement.close();
+            } catch (SQLException e) {
+                throw new DaoException(e.getMessage());
+            }
+        }
+        baseDao.closeConnection();
+    }
+
+    @Override
     public void remove(Client client) throws DaoException {
         Connection conn = baseDao.openConnection();
-        String sql = "delete from clients where id = (?)";
+        String sql = "update clients set id_active_account = null where id = (?)";
         try {
+            final PreparedStatement deleteActiveAccountPreparedStatement = conn.prepareStatement(sql);
+            deleteActiveAccountPreparedStatement.setLong(1, client.getId());
+            if (deleteActiveAccountPreparedStatement.executeUpdate() == 0) {
+                throw new DaoException("impossible to remove the client in db. transaction is rolled back");
+            }
+            deleteActiveAccountPreparedStatement.close();
+            AccountDao accountDao = new AccountDaoImpl();
+            accountDao.removeByClientId(client.getId());
+            sql = "delete from clients where id = (?)";
             final PreparedStatement preparedStatement = conn.prepareStatement(sql);
             preparedStatement.setLong(1, client.getId());
 
-            if(preparedStatement.executeUpdate() == 0){
+            if (preparedStatement.executeUpdate() == 0) {
                 throw new DaoException("impossible to remove the client in db. transaction is rolled back");
             }
             preparedStatement.close();
